binary testing

+++ scripts.py
@@ -838,7 +838,7 @@ def bin_convert_arr(img_arr, val):
     for i in range(img_arr.shape[0]):
         res = np.zeros_like(img_arr[i])
         res[img_arr[i]>val] = 1
-        res = res.astype(np.uint8)
+        res = res.astype(np.float)
         res_list.append(res)
     return np.asarray(res_list)   
 def multi_bin_convert_list(imgList,vals, conv_type = np.int32):
